Todas as funcionalidades de carrinhos adicionadas nesse commit

+++ app/src/main/java/com/projeto/livrariadigitaleclb/data/local/entity/Venda.java
@@ -9,9 +9,9 @@ public class Venda {
     @PrimaryKey(autoGenerate = true)
     public int id;
 
-    public int livroId;
+    public long livroId;  // Mudei de int para long para compatibilidade
     public String titulo;
     public double preco;
 
-    public long dataVenda;
+    public long dataVenda;  // Timestamp em millisegundos
 }
